requirements/jenkins.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/jenkins.txt requirements/base.txt requirements/dev.txt requirements/jenkins.in
#
astroid==1.6.6
beautifulsoup4==4.8.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cssselect==1.0.3
django-admin-index==1.0.1
django-appconf==1.0.2
django-axes==4.4.0
django-better-admin-arrayfield==1.0.2
django-choices==1.7.0
django-cors-middleware==1.3.1
django-debug-toolbar==2.0
django-extra-fields==1.2.3
django-filter==2.0.0
django-ipware==2.1.0
django-jenkins==0.110.0
django-markup==1.3
django-ordered-model==3.3.0
django-privates==1.0.1.post0
django-relativedelta==1.0.5
django-rosetta==0.9.3
django-sendfile2==0.4.2
django-solo==1.1.3
django-webtest==1.9.7
django==2.2.2
djangorestframework-camel-case==0.2.0
djangorestframework-gis==0.14
djangorestframework==3.9.4
drf-flex-fields==0.5.0
drf-nested-routers==0.90.2
drf-writable-nested==0.4.3
drf-yasg==1.16.0
factory-boy==2.12.0
faker==2.0.0
first==2.0.1
freezegun==0.3.12
gemma-zds-client==0.11.0
humanize==0.5.1
idna==2.6
inflection==0.3.1
iso-639==0.4.5
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.4.1
lxml==4.4.0
markdown==3.0.1
markupsafe==1.0
mccabe==0.6.1
oyaml==0.7
pep8==1.7.1
pillow==6.1.0
pip-tools==2.0.2
polib==1.1.0
psycopg2-binary==2.7.5
pyjwt==1.6.4
pylint==1.9.5
pyquery==1.4.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2019.1
pyyaml==3.12
raven==6.9.0
requests-mock==1.6.0
requests==2.21.0
ruamel.yaml==0.15.37
six==1.11.0
soupsieve==1.9.2
sqlparse==0.3.0
tblib==1.4.0
text-unidecode==1.2
unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.24.3
vng-api-common==1.0.7
waitress==1.3.0
webob==1.8.5
webtest==2.0.33
wrapt==1.11.2
